// backend/smart-contracts/typechain-types/VehicleManagement.ts
export declare namespace VehicleManagement {
  export type VehicleStruct = {
    vehicleId: BigNumberish;
    owner: AddressLike;
    model: BytesLike;
    purchaseDate: BigNumberish;
    vin: BytesLike;
    color: BytesLike;
    plateNumber: BytesLike;
    createdAt: BigNumberish;
    updatedAt: BigNumberish;
  };

  export type VehicleStructOutput = [
    vehicleId: bigint,
    owner: string,
    model: string,
    purchaseDate: bigint,
    vin: string,
    color: string,
    plateNumber: string,
    createdAt: bigint,
    updatedAt: bigint
  ] & {
    vehicleId: bigint;
    owner: string;
    model: string;
    purchaseDate: bigint;
    vin: string;
    color: string;
    plateNumber: string;
    createdAt: bigint;
    updatedAt: bigint;
  };
}